+++ index.d.ts
@@ -1,2 +1,6 @@
 export { default as Button } from './types/Button';
 export { default as Icon } from './types/Icon';
+export { default as Breadcrumbs } from './types/Breadcrumbs';
+export { default as Paragraph } from './types/Paragraph';
+export { default as Title } from './types/Title';
+export { default as Spinner } from './types/Spinner';